# Replaced concat with push for better performance per #682