## Description
Replaces instances of `d = d.concat(...)` with `push` within `cmap` functions for improved performance per https://github.com/opentypejs/opentype.js/issues/682. Should not impact results.  Conceptually speaking, this is almost identical to #513.

Additionally, `t.segments.sort` was being run *every time* a new element was added within a loop; I edited to only run once after all elements are inserted. 

## Motivation and Context
See #682.  In short, the performance of `font.toArrayBuffer()` can be poor, including >30+ second runtimes for certain large fonts.  This PR fixes that.

On my system (Chrome, Windows 10), this PR reduces the runtime for writing NotoSansSC from 36 seconds to 8 seconds (78% reduction), and reduces the runtime for NotoSerifSC from 33 seconds to 8 seconds (76% reduction).  

## How Has This Been Tested?
The unit tests pass and I am using a version with these changes in my project. Additionally, the change is conceptually simple and narrow in scope.

## Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist:
<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] I did `npm run test` and all tests passed green (including code styling checks).
- [ ] I have added tests to cover my changes.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the **README** accordingly.
- [x] I have read the **CONTRIBUTING** document.