@Connum 2024-03-22
Could you add a performance test by measuring the time it takes to run `toArrayBuffer()` and making sure it's below a certain threshold that it was above without the change?